# [Enhancement] optimize the performance for max(varchar)