--- a/be/src/util/memcmp.h
+++ b/be/src/util/memcmp.h
@@ -38,6 +38,8 @@
 #include <cstring>
 #include <type_traits>
 
+#include "gutil/strings/fastmem.h"
+
 // Must include headers out of namespace
 #if defined(__SSE4_1__) && !defined(ADDRESS_SANITIZER)
 #include <smmintrin.h>
@@ -152,6 +154,36 @@ inline bool memequal(const char* p1, size_t size1, const char* p2, size_t size2)
 
 inline int memcompare(const char* p1, size_t size1, const char* p2, size_t size2) {
     size_t min_size = std::min(size1, size2);
+    auto res = memcmp(p1, p2, min_size);
+    if (res != 0) {
+        return res > 0 ? 1 : -1;
+    }
+    return compare(size1, size2);
+}
+
+#if defined(__SSE4_2__)
+inline int sse_memcmp2(const char* p1, const char* p2, size_t size) {
+    __m128i left = _mm_lddqu_si128((__m128i*)(p1));
+    __m128i right = _mm_lddqu_si128((__m128i*)(p2));
+    __m128i nz = ~_mm_cmpeq_epi8(left, right);
+    unsigned short mask = _mm_movemask_epi8(nz);
+    volatile int index = __builtin_ctz(mask);
+    if (index >= size) return 0;
+    return (int)(uint8_t)p1[index] - (int)(uint8_t)p2[index];
+}
+#endif
+
+// memcmp has special inline optimizations for bytes <= 16.
+// Requires input to be overflow readable. (Allocate memory aligned to 16 byte size or tail length of 16.)
+inline int memcompare_padded(const char* p1, size_t size1, const char* p2, size_t size2) {
+    size_t min_size = std::min(size1, size2);
+#if defined(__SSE4_2__)
+    if (min_size <= 16) {
+        auto res = sse_memcmp2(p1, p2, min_size);
+        if (res == 0) return compare(size1, size2);
+        return res > 0 ? 1 : -1;
+    }
+#endif
     auto res = memcmp(p1, p2, min_size);
     if (res != 0) {
         return res > 0 ? 1 : -1;