--- a/be/src/util/memcmp.h
+++ b/be/src/util/memcmp.h
@@ -173,12 +173,13 @@ inline int sse_memcmp2(const char* p1, const char* p2, size_t size) {
 }
 #endif
 
+constexpr size_t PADDED_SIZE = 16;
 // memcmp has special inline optimizations for bytes <= 16.
 // Requires input to be overflow readable. (Allocate memory aligned to 16 byte size or tail length of 16.)
 inline int memcompare_padded(const char* p1, size_t size1, const char* p2, size_t size2) {
     size_t min_size = std::min(size1, size2);
 #if defined(__SSE4_2__)
-    if (min_size <= 16) {
+    if (min_size > 0 && min_size <= 16) {
         auto res = sse_memcmp2(p1, p2, min_size);
         if (res == 0) return compare(size1, size2);
         return res > 0 ? 1 : -1;