--- a/be/test/CMakeLists.txt
+++ b/be/test/CMakeLists.txt
@@ -454,6 +454,7 @@ set(EXEC_FILES
         ./util/json_util_test.cpp
         ./util/json_flattener_test.cpp
         ./util/md5_test.cpp
+        ./util/memcmp_test.cpp
         ./util/monotime_test.cpp
         ./util/mysql_row_buffer_test.cpp
         ./util/new_metrics_test.cpp