## Why I'm doing:

1. avoid init std::vector<Slice> in PlainPage::next_batch
2. avoid call virtual call in Max Aggregate Function
3. optimize the memcmp in small strings

baseline: SSB 1BE dop=4 `set cbo_enable_low_cardinality_optimize=false;`
```
   18.29%  libc.so.6         [.] __memcmp_evex_movbe
   18.17%  starrocks_be      [.] starrocks::AggregateFunctionBatchHelper<starrocks::MaxAggregateData<(starrocks::LogicalT
   11.39%  starrocks_be      [.] starrocks::append_fixed_length<unsigned int, 8ul>
    6.29%  starrocks_be      [.] starrocks::BinaryDictPageDecoder<(starrocks::LogicalType)17>::next_batch
    6.10%  starrocks_be      [.] starrocks::BinaryColumnBase<unsignedint>::get
```

```
   mysql>  select max(lo_shipmode) from lineorder;
   +------------------+
   | max(lo_shipmode) |
   +------------------+
   | TRUCK            |
   +------------------+
   1 row in set (1.46 sec)
```

patched:

   avoid virtual function call and avoid init slice contribute about 0.3 sec.
   SSEMemCmp contribute about 0.2 sec.
```
  29.07%  starrocks_be      [.] starrocks::AggregateFunctionBatchHelper<starrocks::MaxAggregateData<(starrocks::LogicalT
  17.35%  starrocks_be      [.] starrocks::append_fixed_length<unsigned int, 8ul>
   8.53%  starrocks_be      [.] starrocks::BinaryPlainPageDecoder<(starrocks::LogicalType)17>::batch_string_at_index
   4.86%  libc.so.6         [.] __memmove_evex_unaligned_erms
```

```
   mysql>  select max(lo_shipmode) from lineorder;
   +------------------+
   | max(lo_shipmode) |
   +------------------+
   | TRUCK            |
   +------------------+
   1 row in set (0.71 sec)
```

benchmark for SSE memcmp
```
   --------------------------------------------------------
   Benchmark              Time             CPU   Iterations
   --------------------------------------------------------
   GlibcCmpImpl        40.2 ns         40.2 ns     17368765
   SSEMemCmpImpl       25.4 ns         25.4 ns     27569981
```

## What I'm doing:

Fixes #issue

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.4
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0